% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compr-package.R
\docType{package}
\name{compr-package}
\alias{compr}
\alias{compr-package}
\title{compr: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\details{
This package is used for caculating compliance costs.
First, the function compliance_data is used to read the data in and convert to the requesite class.
}
\author{
\strong{Maintainer}: Joshua Halls \email{Joshua.Halls@ons.gov.uk}

}
\keyword{internal}
